+++ packages/cactus-common/src/main/typescript/index.ts
@@ -1,1 +1,2 @@
 export * from "./public-api";
+export { IListenOptions, Servers } from "./servers";
